requirements.txt
amqp==2.6.0
appdirs==1.4.4
asgiref==3.2.8
billiard==3.6.3.0
celery==4.4.5
certifi==2020.4.5.2
chardet==3.0.4
distlib==0.3.0
Django==3.0
django-crispy-forms==1.9.1
django-redis==4.12.1
filelock==3.0.12
future==0.18.2
gunicorn==20.0.4
idna==2.9
kombu==4.6.10
pipenv==2020.6.2
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
virtualenv==20.0.23
virtualenv-clone==0.5.4
